requirements.txt
affine==2.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
astroid==4.0.3
attrs==25.4.0
autobahn==25.12.2
autoflake==2.3.1
Automat==25.4.16
babel==2.17.0
billiard==4.2.4
black==26.3.1
boto3==1.42.33
botocore==1.42.33
cachetools==7.0.1
cbor2==5.9.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
cfgv==3.4.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cligj==0.7.2
color-operations==0.2.0
constantly==23.10.4
cryptography==46.0.7
daphne==4.2.1
dill==0.4.1
distlib==0.3.9
distro==1.9.0
Django==6.0.4
django-cors-headers==4.9.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-money==3.5.4
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
filelock==3.20.3
gunicorn==23.0.0
h11==0.16.0
hiredis==3.0.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
identify==2.6.1
idna==3.11
Incremental==24.11.0
isort==7.0.0
jiter==0.14.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.1.1
kombu==5.6.2
langchain==1.2.15
langchain-core==1.3.0
langchain-ollama==1.1.0
langchain-openai==1.1.14
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.31
language_tool_python==3.2.2
Markdown==3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mercantile==1.2.1
minio==7.2.20
morecantile==7.0.3
msgpack==1.1.2
mypy-extensions==1.0.0
nodeenv==1.9.1
numexpr==2.14.1
numpy==2.4.2
ollama==0.6.1
openai==2.32.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==25.0
pathspec==1.0.1
pillow==12.2.0
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.52
psutil==7.2.1
psycopg==3.3.2
psycopg-binary==3.3.2
pubcontrol==3.5.0
py-moneyed==3.0
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.2.0
PyJWT==2.12.0
pylint==4.0.4
pyOpenSSL==26.0.0
pyparsing==3.3.2
pyproj==3.7.2
pystac==1.14.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rasterio==1.5.0
redis==7.1.0
regex==2026.4.4
requests==2.33.0
requests-toolbelt==1.0.0
rio-cogeo==7.0.1
rio-mucho==1.0.0
rio-rgbify==0.4.0
rio-tiler==9.0.6
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
s3transfer==0.16.0
scipy==1.17.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tomlkit==0.14.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
types-PyYAML==6.0.12.20250915
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.40.0
vine==5.1.0
virtualenv==20.36.1
wcwidth==0.6.0
Werkzeug==3.1.6
wheel==0.46.2
wsproto==1.2.0
xxhash==3.6.0
yaml-1.3==0.1.0
zope.interface==8.2
zstandard==0.25.0
geopy==2.4.1
matplotlib==3.10.8
shapely==2.1.2
python-json-logger==4.1.0
